feat: improve main page layout and tests disable
