+++ next.config.ts
@@ -1,9 +1,7 @@
 import createNextIntlPlugin from 'next-intl/plugin';
 import type { NextConfig } from 'next';
 
-const nextConfig: NextConfig = {
-  distDir: './dist',
-};
+const nextConfig: NextConfig = {};
 
 const withNextIntl = createNextIntlPlugin();
 export default withNextIntl(nextConfig);
